% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.p.sig.R
\name{add.n.r}
\alias{add.n.r}
\title{Add row numbers to regression result data.frame}
\usage{
add.n.r(df)
}
\arguments{
\item{df}{a data.frame of regression result}
}
\description{
Add row numbers to regression result data.frame
}
