% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.p.sig.R
\name{add.p.z}
\alias{add.p.z}
\title{Add approximate p.value based on t.value or z.value, when sample size is large}
\usage{
add.p.z(df, z.col = 3, p.already = FALSE)
}
\arguments{
\item{df}{a data.frame of regression result}

\item{z.col}{the column number of t.value or z.value}

\item{p.already}{whether the regression result already contains p.value}
}
\description{
Add approximate p.value based on t.value or z.value, when sample size is large
}
