% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.p.sig.R
\name{add.pr}
\alias{add.pr}
\title{Add approximate p-value based on t score or z score, when sample size is large}
\usage{
add.pr(df, z.col = 3, p.already = FALSE)
}
\arguments{
\item{df}{a data.frame of regression result}

\item{z.col}{the column number of t score or z score}

\item{p.already}{whether the regression result already contains p.value}
}
\description{
Add approximate p-value based on t score or z score, when sample size is large
}
