% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.check.R
\name{check_cor}
\alias{check_cor}
\title{quickly check correlation matrix, or the correlation between a particular X and all other vars 
could be useful for looking for relevant instrument}
\usage{
check_cor(data, var_name_select = NULL, d = 3)
}
\arguments{
\item{data}{a data.frame used in regression model}

\item{var_name_select}{to specify the variable names to be included in the table, default is NULL--all variables are included}

\item{d}{number of digits retained after the decimal point}
}
\description{
quickly check correlation matrix, or the correlation between a particular X and all other vars 
could be useful for looking for relevant instrument
}
\examples{
data(mtcars)
check_cor(mtcars)

}
