% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.check.R
\name{check_vif}
\alias{check_vif}
\title{quickly check the vifs in a regression model; for checking multi-collinearity}
\usage{
check_vif(data)
}
\arguments{
\item{data}{a data.frame used in regression model}
}
\description{
quickly check the vifs in a regression model; for checking multi-collinearity
}
\examples{
data(mtcars)
model <- lm(mpg ~ vs + carb + hp + wt + wt * hp , data = mtcars)
check_vif(data = model$model)

}
