% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.testing.R
\name{coef_equality}
\alias{coef_equality}
\title{testing equality of two coefficients, a Wald test}
\usage{
coef_equality(model, var1.name, var2.name, v = NULL)
}
\arguments{
\item{model}{the model object, such as a "lm" object}

\item{var1.name}{X1 name in model, a string}

\item{var2.name}{X2 name in model, a string}

\item{v}{a customized variance-covariance matrix}
}
\description{
testing equality of two coefficients, a Wald test
}
