% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.p.sig.R
\name{format_reg.table}
\alias{format_reg.table}
\title{Convert the regression result to the long format: the standard errors are in parentheses and beneath the betas}
\usage{
format_reg.table(df, d = 3)
}
\arguments{
\item{df}{a data.frame of regression result}

\item{d}{number of decimal places to retain}
}
\description{
Convert the regression result to the long format: the standard errors are in parentheses and beneath the betas
}
