% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.packages.R
\name{load.pkgs}
\alias{load.pkgs}
\title{load multiple packages}
\usage{
load.pkgs(pkg_name_vec)
}
\arguments{
\item{pkg_name_vec}{a string vector of package names}
}
\description{
load multiple packages
}
