% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.table.R
\name{reg.combine}
\alias{reg.combine}
\title{Combine regression results from different models by columns # updated 9/2/2018 #}
\usage{
reg.combine(tbl_1, tbl_2, tbl_3 = NULL, tbl_4 = NULL, tbl_5 = NULL,
  tbl_6 = NULL, tbl_7 = NULL, tbl_8 = NULL, tbl_9 = NULL,
  tbl_10 = NULL, tbl_11 = NULL, tbl_12 = NULL, tbl_13 = NULL,
  tbl_14 = NULL, tbl_15 = NULL, tbl_16 = NULL, tbl_17 = NULL,
  tbl_18 = NULL, tbl_19 = NULL, tbl_20 = NULL)
}
\arguments{
\item{tbl_1}{the 1st  data.frame of regression result}

\item{tbl_2}{the 2nd  data.frame of regression result}

\item{tbl_3}{the 3rd  data.frame of regression result}

\item{tbl_4}{the 4th  data.frame of regression result}

\item{tbl_5}{the 5th  data.frame of regression result}

\item{tbl_6}{the 6th  data.frame of regression result}

\item{tbl_7}{the 7th  data.frame of regression result}

\item{tbl_8}{the 8th  data.frame of regression result}

\item{tbl_9}{the 9th  data.frame of regression result}

\item{tbl_10}{the 10th data.frame of regression result}
}
\description{
Combine regression results from different models by columns # updated 9/2/2018 #
}
