% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.check.R
\name{reg.Cor}
\alias{reg.Cor}
\title{quickly check correlation matrix, or the correlation between a particular X and all other vars 
can be useful for looking for relevant instrument}
\usage{
reg.Cor(model_df, var_name_select = NULL, d = 3)
}
\arguments{
\item{model_df}{a data.frame used in regression model}

\item{var_name_select}{to specify the variable names to be included in the table, default is NULL--all variables are included}

\item{d}{number of digits retained after the decimal point}
}
\description{
quickly check correlation matrix, or the correlation between a particular X and all other vars 
can be useful for looking for relevant instrument
}
