% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.table.R
\name{reg.Cor.Table}
\alias{reg.Cor.Table}
\title{make the correlation matrix from the data.frame used in regression}
\usage{
reg.Cor.Table(model_df, model_df_to_combine = NULL,
  var_name_select = NULL, all.var.names = NULL, d = 2)
}
\arguments{
\item{model_df}{a data.frame used in regression model, e.g. model$model}

\item{model_df_to_combine}{another data.frame used for regression model, e.g. when you have similar set of X's but different Y's}

\item{var_name_select}{specify the names of regression variables to be included in the correlation matrix}

\item{all.var.names}{all variable names, a string vector}

\item{d}{number of decimal places to retain}
}
\description{
make the correlation matrix from the data.frame used in regression
}
