% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggploting_simple_interaction.R
\name{reg.gg}
\alias{reg.gg}
\title{plotting the interaction, with or without multiple interaction terms}
\usage{
reg.gg(reg.result, df, by_color = FALSE, x_var.name = NULL,
  y_var.name = NULL, main1.r, mdrt.r = NULL, int1.r = NULL,
  min_x = 0.001, max_x = 0.999, mdrt_quantile_05 = NULL,
  mdrt_quantile_50 = NULL, mdrt_quantile_95 = NULL, mod.n.sd = NULL,
  title = NULL, xlab = "X_Var.name", ylab = "Y_Var.name",
  moderator.lab = "Moderator_name", mdrt.low.name = "Low",
  mdrt.mid.name = NULL, mdrt.high.name = "High", y.hi.lim = NULL,
  y.low.lim = NULL)
}
\arguments{
\item{reg.result}{a data.frame of regression result}

\item{df}{the data used in regression}

\item{by_color}{plot interactions by colors, otherwise by line types}

\item{x_var.name}{x name in the regression model, a string}

\item{y_var.name}{y name in the regression model, a string}

\item{main1.r}{the row number of the main effect, in reg.result}

\item{mdrt.r}{the row number of the moderater, in reg.result}

\item{int1.r}{the row number of the interaction, in reg.result}

\item{min_x}{the min of x scale, in percentile of x}

\item{max_x}{the max of x scale, in percentile of x}

\item{mdrt_quantile_05}{set the low level of moderator, in percentile}

\item{mdrt_quantile_50}{set the middle level of moderator, in percentile}

\item{mdrt_quantile_95}{set the high level of moderator, in percentile}

\item{mod.n.sd}{set the moderating strength, in the number of s.d. units, which can take negative values}

\item{title}{the title of the plot}

\item{xlab}{label of X}

\item{ylab}{label of Y}

\item{moderator.lab}{label of moderator}

\item{mdrt.low.name}{the label of low-level moderator}

\item{mdrt.mid.name}{the label of mid-level moderator}

\item{mdrt.high.name}{the label of high-level moderator}

\item{y.hi.lim}{specify the upper limit of y}

\item{y.low.lim}{specify the lower limit of y}
}
\description{
plotting the interaction, with or without multiple interaction terms
}
