% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.table.R
\name{reg.table}
\alias{reg.table}
\title{Convert a regression result into a clean data.frame}
\usage{
reg.table(lm.obj)
}
\arguments{
\item{lm.obj}{a regression result object}
}
\description{
Convert a regression result into a clean data.frame
}
