% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.check.R
\name{reg.Vif}
\alias{reg.Vif}
\title{quickly check the vifs in a regression model; for checking multi-collinearity}
\usage{
reg.Vif(model_df)
}
\arguments{
\item{model_df}{a data.frame used in regression model}
}
\description{
quickly check the vifs in a regression model; for checking multi-collinearity
}
