% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.testing.R
\name{slope.sig_after.mod}
\alias{slope.sig_after.mod}
\title{significance of slope under moderation
testing restriction: the sig. of beta_x under the moderation of z1, with or without additional interaction terms, Aug 13th}
\usage{
slope.sig_after.mod(m, model, mod_name, mod.n.sd = 1, full.data)
}
\arguments{
\item{m}{a data.frame of regression result}

\item{model}{the model object, such as a "lm" object}

\item{mod_name}{moderator name in model, a string}

\item{mod.n.sd}{specify the strength of the moderating effects, in the unit of s.d.s of the moderator, which can take negative values}

\item{full.data}{data used for regression}
}
\description{
significance of slope under moderation
testing restriction: the sig. of beta_x under the moderation of z1, with or without additional interaction terms, Aug 13th
}
