% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.p.sig.R
\name{add.sig}
\alias{add.sig}
\title{Add significance marks to the regression result}
\usage{
add.sig(df, Pr.col = 5)
}
\arguments{
\item{df}{a data.frame of regression result}

\item{Pr.col}{the column number of p.value}
}
\description{
Add significance marks to the regression result
}
