% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.check.R
\name{check_na_in}
\alias{check_na_in}
\title{quickly check the proportion of NAs in each columns of a dataframe}
\usage{
check_na_in(df, true_total = FALSE)
}
\arguments{
\item{df}{a data.frame}

\item{true_total}{FALSE to show the percentage, TRUE to show the true number of missing values}
}
\description{
quickly check the proportion of NAs in each columns of a dataframe
}
