% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrrr.R
\docType{package}
\name{regrrr}
\alias{regrrr}
\alias{regrrr-package}
\title{regrrr: a toolkit for compiling regression results}
\description{
Compiling, Testing, Plotting Regression Results
}
\author{
R.K. Yang \email{rkzyang@gmail.com}

Luyao Peng \email{luyaopeng.cn@gmail.com}
}
